controller almost separated from application
