+++ src/Controller.h
@@ -14,12 +14,25 @@
 #include "IO/AudioController.h"
 #include "helpers/ErrorReport.h"
 #include "imgui.h"
-#include "Application/AppSettings.h"
+#include "application/AppSettings.h"
 #include "Gui.h"
-#include "Application/Keybinds.h"
+#include "application/Keybinds.h"
 #include "View.h"
 
 class Controller {
+public:
+
+
+    Controller(const std::shared_ptr<AppSettings> &settings, const std::shared_ptr<View> &view,const std::shared_ptr<GameBoy> &gameBoy);
+    void handleSDLEvents();
+
+private:
+    void handleEmulatorInputPress(SDL_Keycode key);
+    void handleEmulatorInputRelease(SDL_Keycode key);
+    std::shared_ptr<AppSettings> settings;
+    std::shared_ptr<View> view;
+    std::shared_ptr<GameBoy> gameBoy;
+    float savedEmulationSpeed;
 
 };
